IMPORTANT

+++ private/controllers/Classes.php
@@ -17,7 +17,44 @@ class Classes extends Controller
 
 		$classes = new Classes_model();
 
-		$data = $classes->findAll();
+		$school_id = Auth::getSchool_id();
+		
+		if(Auth::access('admin')){
+
+			$query = "select * from classes where school_id = :school_id && year(date) = :school_year order by id desc";
+			$arr['school_id'] = $school_id;
+			$arr['school_year'] = !empty($_SESSION['SCHOOL_YEAR']->year) ? $_SESSION['SCHOOL_YEAR']->year : date("Y",time());
+
+			if(isset($_GET['find']))
+	 		{
+	 			$find = '%' . $_GET['find'] . '%';
+	 			$query = "select * from classes where school_id = :school_id && (class like :find) && year(date) = :school_year order by id desc";
+	 			$arr['find'] = $find;
+	 		}
+
+			$data = $classes->query($query,$arr);
+ 		}else{
+
+ 			$class = new Classes_model();
+ 			$mytable = "class_students";
+ 			if(Auth::getRank() == "lecturer"){
+ 				$mytable = "class_lecturers";
+ 			}
+ 			
+			$query = "select * from classes where (class_id in (select class_id from $mytable where user_id = :user_id && disabled = 0) && year(date) = :school_year) || user_id = :user_id ";
+ 			$arr['user_id'] = Auth::getUser_id();
+ 			$arr['school_year'] = !empty($_SESSION['SCHOOL_YEAR']->year) ? $_SESSION['SCHOOL_YEAR']->year : date("Y",time());
+
+ 			if(isset($_GET['find']))
+	 		{
+	 			$find = '%' . $_GET['find'] . '%';
+	 			$query = "select classes.class, {$mytable}.* from $mytable join classes on classes.class_id = {$mytable}.class_id where ({$mytable}.user_id = :user_id && {$mytable}.disabled = 0 && classes.class like :find && year(classes.date) = :school_year ) ";
+	 			$arr['find'] = $find;
+	 		}
+
+ 			$data = $class->query($query,$arr);
+ 
+ 		}
 
 		$crumbs[] = ['Dashboard',''];
 		$crumbs[] = ['Classes','classes'];
@@ -75,9 +112,10 @@ class Classes extends Controller
 		}
 
 		$classes = new Classes_model();
+ 		$row = $classes->where('id',$id);
 
 		$errors = array();
-		if(count($_POST) > 0)
+		if(count($_POST) > 0 && Auth::access('lecturer') && Auth::i_own_content($row))
  		{
 
 			if($classes->validate($_POST))
@@ -92,17 +130,21 @@ class Classes extends Controller
  			}
  		}
 
- 		$row = $classes->where('id',$id);
 
  		$crumbs[] = ['Dashboard',''];
 		$crumbs[] = ['Classes','classes'];
 		$crumbs[] = ['Edit','classes/edit'];
 
-		$this->view('classes.edit',[
-			'row'=>$row,
-			'errors'=>$errors,
-			'crumbs'=>$crumbs,
-		]);
+		if(Auth::access('lecturer') && Auth::i_own_content($row)){
+
+			$this->view('classes.edit',[
+				'row'=>$row,
+				'errors'=>$errors,
+				'crumbs'=>$crumbs,
+			]);
+		}else{
+			$this->view('access-denied');
+		}
 	}
 
 	public function delete($id = null)
@@ -113,11 +155,13 @@ class Classes extends Controller
 			$this->redirect('login');
 		}
 
+ 
 		$classes = new Classes_model();
+ 		$row = $classes->where('id',$id);
 
 		$errors = array();
 
-		if(count($_POST) > 0)
+		if(count($_POST) > 0 && Auth::access('lecturer') && Auth::i_own_content($row))
  		{
  
  			$classes->delete($id);
@@ -125,16 +169,20 @@ class Classes extends Controller
  		 
  		}
 
- 		$row = $classes->where('id',$id);
 
  		$crumbs[] = ['Dashboard',''];
 		$crumbs[] = ['Classes','classes'];
 		$crumbs[] = ['Delete','classes/delete'];
 
-		$this->view('classes.delete',[
-			'row'=>$row,
- 			'crumbs'=>$crumbs,
-		]);
+		if(Auth::access('lecturer') && Auth::i_own_content($row)){
+
+			$this->view('classes.delete',[
+				'row'=>$row,
+	 			'crumbs'=>$crumbs,
+			]);
+		}else{
+			$this->view('access-denied');
+		}
 	}
 	
 }
